moved response headers to better place
