+++ src/libs/http/headers.ts
@@ -1,0 +1,5 @@
+// This file contains the headers used in HTTP requests and responses.
+export const responseHeaders = {
+  "Access-Control-Allow-Origin": "*",
+  "Access-Control-Allow-Credentials": true
+};
